+++ packages/elysia/src/createAuth.ts
@@ -1,13 +1,15 @@
 import { createAuth as createAuthBasic, type AuthConfig } from "@aura-stack/auth"
 import { toHandler } from "@/lib/handler"
 import { withAuth } from "@/lib/with-auth"
-import type { FromShapeToObject, Identities } from "@aura-stack/auth/identity"
+import type { FromShapeToObject, Identities, SchemaTypes } from "@aura-stack/auth/identity"
 
 /**
  * Creates an Aura Auth instance with Elysia-specific utilities.
  */
-export const createAuth = <Identity extends Identities>(config: AuthConfig<Identity>) => {
-    const auth = createAuthBasic<Identity>(config)
+export const createAuth = <Identity extends Identities, SignUpSchema extends SchemaTypes>(
+    config: AuthConfig<Identity, SignUpSchema>
+) => {
+    const auth = createAuthBasic<Identity, SignUpSchema>(config)
 
     return {
         ...auth,