+++ packages/hono/src/createAuth.ts
@@ -2,10 +2,12 @@ import { createAuth as createAuthBasic, type AuthConfig } from "@aura-stack/auth
 import { toHandler } from "@/lib/handler"
 import { withAuth } from "@/lib/with-auth"
 import type { Context } from "hono"
-import type { Identities, FromShapeToObject } from "@aura-stack/auth/identity"
+import type { Identities, FromShapeToObject, SchemaTypes } from "@aura-stack/auth/identity"
 
-export const createAuth = <Identity extends Identities>(config: AuthConfig<Identity>) => {
-    const auth = createAuthBasic<Identity>(config)
+export const createAuth = <Identity extends Identities, SignUpSchema extends SchemaTypes>(
+    config: AuthConfig<Identity, SignUpSchema>
+) => {
+    const auth = createAuthBasic<Identity, SignUpSchema>(config)
 
     return {
         ...auth,